Update NINJA PIRATES BOSS ROOM.xml
Add visual cues corresponding to CRE blocks in the middle screen of the vanilla room. The two groups of three tiles on the floor are where to stand to make the pirates jump and become vulnerable. The three single blocks on the ceiling are where the "screens" hang down in the vanilla room, a visual cue for killing the pirates with shinespark echoes