Mark unused tiles with X's
Looks like we're going to need to add quite a few new tiles for this tileset, and it looked like there wasn't much space. But there were a ton of garbage tiles (16x16 and 8x8) that aren't used, so I overwrote them with X's so we can keep track that they are free.