NorfairPalette crateria tweaks
This tones down the orange highlight colors which did not seem to blend in well in the Crateria tilesets. Also adjusts the relationship between inner crateria awake vs. asleep palettes to make it more like vanilla, in particular paying attention to the Climb platforms which are supposed to have some lighter shades in the asleep palette, helping them be more visible in the fog.