Update CRATERIA GAUNTLET ENTRANCE.xml
Remove fake slope tiles around Yapping Maws. I assume these were trying to imitate the bug "nest" from vanilla, but I think the effect doesn't really work in this theme? To me it didn't look great, because of how it didn't connect properly with the surrounding bricks. And it seemed to misrepresent the collision