Update WRECKED SHIP ENTRANCE.xml
Remove the metal support foreground decorations. The bottom row gets deleted by the game because of the treadmills, so it looked broken. Also they didn't really look right, with how the left and right halves came together to a point instead of a proper junction. Decorations don't really seem necessary in this space anyway?